Basic wall jumping implementation
- I moved the camera and slightly adjusted the terrain in order to test
this.
- Code at the moment checks if a player in midair is colliding with
terrain at the sides (and only at the sides) - if so, they can jump
without it being added to their jump limit.
- At the moment this means they can just "climb" a wall so to speak.
This will have to be fixed as some point, this is only a bare bones
implementation.
- I cleaned up a bit of dodgy code. I will also be fixing the get/set
functions already in there, as they are ghastly. There's special C#
syntax that we should use as it makes everything a whole lot easier.